# [perf] Precompile for instantiate